Put Reader back in state
