+++ src/parsita/options.py
@@ -11,9 +11,8 @@ __all__ = [
 import re
 from typing import Any, Sequence
 
-from .reader import SequenceReader, StringReader
 from .result import Result
-from .state import Input, Output
+from .state import Input, Output, SequenceReader, StringReader
 
 # Global mutable state
