+++ src/parsita/parsers/_alternative.py
@@ -2,8 +2,7 @@ __all__ = ["AlternativeParser", "first", "LongestAlternativeParser", "longest"]
 
 from typing import Generic, Optional, Sequence, Union
 
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
 from ._literal import lit
