+++ src/parsita/parsers/_conversion.py
@@ -2,8 +2,7 @@ __all__ = ["ConversionParser", "TransformationParser"]
 
 from typing import Callable, Generic, Optional, TypeVar
 
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
 
 Convert = TypeVar("Convert")
