+++ src/parsita/parsers/_optional.py
@@ -2,8 +2,7 @@ __all__ = ["OptionalParser", "opt"]
 
 from typing import Generic, List, Sequence, Union
 
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
 from ._literal import lit
