+++ src/parsita/parsers/_predicate.py
@@ -2,8 +2,7 @@ __all__ = ["PredicateParser", "pred"]
 
 from typing import Callable, Generic
 
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
