+++ src/parsita/parsers/_repeated_seperated.py
@@ -3,8 +3,7 @@ __all__ = ["RepeatedSeparatedParser", "repsep", "RepeatedOnceSeparatedParser", "
 from typing import Any, Generic, Optional, Sequence, Union
 
 from ..exceptions import RecursionError
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
 from ._literal import lit
