+++ src/parsita/parsers/_sequential.py
@@ -2,8 +2,7 @@ __all__ = ["SequentialParser", "DiscardLeftParser", "DiscardRightParser"]
 
 from typing import Any, Generic, List
 
-from ..reader import Reader
-from ..state import Continue, Input, Output, State
+from ..state import Continue, Input, Output, Reader, State
 from ._base import Parser
